+++ frontend/appflowy_flutter/lib/plugins/database/grid/presentation/widgets/filter/choicechip/date.dart
@@ -6,7 +6,7 @@ import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/workspace/presentation/widgets/pop_up_action.dart';
 import 'package:appflowy_backend/protobuf/flowy-database2/protobuf.dart';
-import 'package:appflowy/workspace/presentation/widgets/date_picker/appflowy_date_picker.dart';
+import 'package:appflowy/workspace/presentation/widgets/date_picker/desktop_date_picker.dart';
 import 'package:flowy_infra/size.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
@@ -198,7 +198,7 @@ class _DateFilterEditorState extends State<DateFilterEditor> {
           child: SingleFilterBlocSelector<DateTimeFilter>(
             filterId: widget.filterId,
             builder: (context, filter, field) {
-              return AppFlowyDatePicker(
+              return DesktopAppFlowyDatePicker(
                 isRange: isRange,
                 includeTime: false,
                 dateFormat: DateFormatPB.Friendly,
