+++ frontend/appflowy_flutter/lib/workspace/presentation/widgets/date_picker/widgets/date_time_text_field.dart
@@ -7,7 +7,7 @@ import 'package:flowy_infra/size.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import '../appflowy_date_picker.dart';
+import '../desktop_date_picker.dart';
 import 'date_picker.dart';
 
 class DateTimeTextField extends StatefulWidget {
